Move affine types from ir.py to affine_ir.py
